python/agents/dynamic_workflow-code_reviewer/pyproject.toml
[project]
name = "dynamic_workflow-code_reviewer"
version = "0.1.0"
description = "Takes in a code and reviews it's linting and tests."
authors = [{ name = "Tommy Wagner", email = "wagnerthomas@google.com" }]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.4.6",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
known-first-party = ["dynamic_workflow-code-reviewer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"